issue 99: added feature + fixed tests + reformatted using isort and black
