+++ sporkfish/searcher/negamax_lazy_smp.py
@@ -8,6 +8,7 @@ from sporkfish.board.board import Board
 from sporkfish.evaluator.evaluator import Evaluator
 from sporkfish.searcher.negamax import NegamaxSp
 from sporkfish.searcher.searcher_config import SearcherConfig
+from sporkfish.statistics import NodeTypes, Statistics
 
 
 # This doesn't really work yet. Don't use.
@@ -46,6 +47,7 @@ class NegaMaxLazySmp(NegamaxSp):
         """
 
         def task() -> Tuple[float, chess.Move]:
+            # TODO: fix increment statistics
             return NegamaxSp._start_search_from_root(self, board, depth, alpha, beta)
 
         # Let processes race down lazily and see who completes first
