+++ tests/test_statistics.py
@@ -1,6 +1,7 @@
 import pytest
 
-from sporkfish.statistics import Statistics, NodeTypes
+from sporkfish.statistics import NodeTypes, Statistics
+
 
 class TestIncrementStatistics:
     def test_increment_node_visited(self):
@@ -8,7 +9,7 @@ class TestIncrementStatistics:
         for k, v in S.nodes_visited.items():
             assert v == 0
             S.increment_node_visited(k)
-        
+
         for k, v in S.nodes_visited.items():
             assert v == 1
 
@@ -20,9 +21,10 @@ class TestIncrementStatistics:
 
     def test_increment_nodes_in_tt(self):
         S = Statistics()
-        assert S.nodes_in_tt == 0
-        S.increment_nodes_in_tt()
-        assert S.nodes_in_tt == 1
+        assert S.nodes_from_tt == 0
+        S.increment_nodes_from_tt()
+        assert S.nodes_from_tt == 1
+
 
 class TestResetStatistics:
     def test_reset_node_visited(self):
@@ -31,7 +33,7 @@ class TestResetStatistics:
             assert v == 0
             S.increment_node_visited(k, 3)
             assert S.nodes_visited[k] == 3
-        
+
         S.reset_node_visited()
         for k, v in S.nodes_visited.items():
             assert v == 0
@@ -46,8 +48,8 @@ class TestResetStatistics:
 
     def test_reset_nodes_in_tt(self):
         S = Statistics()
-        assert S.nodes_in_tt == 0
-        S.increment_nodes_in_tt(32)
-        assert S.nodes_in_tt == 32
-        S.reset_nodes_in_tt(0)
-        assert S.nodes_in_tt == 0
+        assert S.nodes_from_tt == 0
+        S.increment_nodes_from_tt(32)
+        assert S.nodes_from_tt == 32
+        S.reset_nodes_from_tt(0)
+        assert S.nodes_from_tt == 0
